+++ packages/iov-bcp/src/connection.ts
@@ -213,20 +213,22 @@ export interface BlockHeader {
   readonly transactionCount: number;
 }
 
-// BcpConnection is a high-level interface to a blockchain node,
-// abstracted over all blockchain types and communication channel.
-// A direct connection or a proxy server should implement this.
-// The implementation takes care to convert our internal types into
-// the proper format for the blockchain.
-//
-// BcpConnection is the minimal interface needed to be supported by any blockchain
-// that is compatible with the bcp spec and iov-core library. This supports
-// getting account balances, sending tokens, and observing the blockchain state.
-//
-// There are other optional interfaces that extend this functionality with
-// features like atomic swap, NFTs, etc which may be implemented by any connector
-// to enable enhanced features in the clients
-export interface BcpConnection {
+/**
+ * A high-level interface to a blockchain node,
+ * abstracted over all blockchain types and communication channel.
+ * A direct connection or a proxy server should implement this.
+ * The implementation takes care to convert our internal types into
+ * the proper format for the blockchain.
+ *
+ * BlockchainConnection is the minimal interface needed to be supported by any blockchain
+ * that is compatible with the BCP spec and IOV-Core library. This supports
+ * getting account balances, sending tokens, and observing the blockchain state.
+ *
+ * There are other optional interfaces that extend this functionality with
+ * features like atomic swap, NFTs, etc which may be implemented by any connector
+ * to enable enhanced features in the clients.
+ */
+export interface BlockchainConnection {
   // blockchain
   readonly disconnect: () => void;
   readonly chainId: () => ChainId;