+++ packages/iov-lisk/src/liskconnection.ts
@@ -8,8 +8,8 @@ import {
   AccountQuery,
   AddressQuery,
   Algorithm,
-  BcpConnection,
   BcpTicker,
+  BlockchainConnection,
   BlockHeader,
   BlockId,
   BlockInfo,
@@ -67,7 +67,7 @@ async function loadChainId(baseUrl: string): Promise<ChainId> {
   return responseBody.data.nethash;
 }
 
-export class LiskConnection implements BcpConnection {
+export class LiskConnection implements BlockchainConnection {
   public static async establish(baseUrl: string): Promise<LiskConnection> {
     const chainId = await loadChainId(baseUrl);
     return new LiskConnection(baseUrl, chainId);